LayoutTests/fast/events/touch/document-create-touch-list-expected.txt

This tests support for the document.createTouchList API.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS "createTouchList" in document is true
PASS touchList is non-null.
PASS touchList.length is 0
PASS touchList.item(0) is null
PASS touchList.item(1) is null
PASS touchList.item() threw exception TypeError: Not enough arguments.
PASS ts.touches.length is 2
PASS ts.touches[0].identifier is 12341
PASS ts.touches[0].clientX is 60
PASS ts.touches[1].screenY is 120
PASS ts.ctrlKey is true
PASS successfullyParsed is true

TEST COMPLETE
